+++ src/renderer_raytrace_clusters_lod.cpp
@@ -56,6 +56,7 @@ private:
     shaderc::SpvCompilationResult computeTraversalPresort;
     shaderc::SpvCompilationResult computeTraversalInit;
     shaderc::SpvCompilationResult computeTraversalRun;
+    shaderc::SpvCompilationResult computeTraversalGroups;
     shaderc::SpvCompilationResult computeBuildSetup;
 
     shaderc::SpvCompilationResult computeBlasInsertClusters;
@@ -73,6 +74,7 @@ private:
     VkPipeline computeTraversalPresort = nullptr;
     VkPipeline computeTraversalInit    = nullptr;
     VkPipeline computeTraversalRun     = nullptr;
+    VkPipeline computeTraversalGroups  = nullptr;
     VkPipeline computeBuildSetup       = nullptr;
 
     VkPipeline computeBlasInsertClusters  = nullptr;
@@ -130,6 +132,7 @@ bool RendererRayTraceClustersLod::initShaders(Resources& res, RenderScene& rscen
   options.AddMacroDefinition("USE_CULLING", config.useCulling ? "1" : "0");
   options.AddMacroDefinition("USE_BLAS_SHARING", config.useBlasSharing ? "1" : "0");
   options.AddMacroDefinition("USE_RENDER_STATS", config.useRenderStats ? "1" : "0");
+  options.AddMacroDefinition("USE_SEPARATE_GROUPS", config.useSeparateGroups ? "1" : "0");
   options.AddMacroDefinition("DEBUG_VISUALIZATION", config.useDebugVisualization ? "1" : "0");
 
   shaderc::CompileOptions optionsAO = options;
@@ -153,6 +156,12 @@ bool RendererRayTraceClustersLod::initShaders(Resources& res, RenderScene& rscen
   else
   {
     res.compileShader(m_shaders.computeTraversalInit, VK_SHADER_STAGE_COMPUTE_BIT, "traversal_init.comp.glsl", &options);
+  }
+
+  if(m_config.useSeparateGroups)
+  {
+    res.compileShader(m_shaders.computeTraversalGroups, VK_SHADER_STAGE_COMPUTE_BIT,
+                      "traversal_run_separate_groups.comp.glsl", &options);
   }
 
   res.compileShader(m_shaders.computeTraversalRun, VK_SHADER_STAGE_COMPUTE_BIT, "traversal_run.comp.glsl", &options);
@@ -247,6 +256,9 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
     m_sceneBuildShaderio.tlasInstances      = m_tlasInstancesBuffer.address;
     m_sceneBuildShaderio.numGeometries      = uint32_t(rscene.scene->getActiveGeometryCount());
 
+    m_sceneBuildShaderio.indirectDispatchGroups.gridY = 1;
+    m_sceneBuildShaderio.indirectDispatchGroups.gridZ = 1;
+
     BufferRanges mem = {};
     m_sceneBuildShaderio.renderClusterInfos =
         mem.append(sizeof(shaderio::ClusterInfo) * m_sceneBuildShaderio.maxRenderClusters, 8);
@@ -268,6 +280,11 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
       m_sceneBuildShaderio.instanceSortKeys   = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
       m_sceneBuildShaderio.instanceSortValues = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
       mem.splitOverlap();
+    }
+
+    if(config.useSeparateGroups)
+    {
+      m_sceneBuildShaderio.traversalGroupInfos = mem.append(sizeof(uint64_t) * m_sceneBuildShaderio.maxTraversalInfos, 8);
     }
 
     m_sceneBuildShaderio.blasBuildSizes     = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
@@ -294,7 +311,10 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
     m_sceneBuildShaderio.instanceSortValues += m_sceneDataBuffer.address;
     m_sceneBuildShaderio.instanceBuildInfos += m_sceneDataBuffer.address;
     m_sceneBuildShaderio.geometryBuildInfos += m_sceneDataBuffer.address;
-
+    if(config.useSeparateGroups)
+    {
+      m_sceneBuildShaderio.traversalGroupInfos += m_sceneDataBuffer.address;
+    }
 
     res.createBuffer(m_sceneTraversalBuffer, sizeof(uint64_t) * m_sceneBuildShaderio.maxTraversalInfos,
                      VK_BUFFER_USAGE_STORAGE_BUFFER_BIT);
@@ -412,6 +432,12 @@ bool RendererRayTraceClustersLod::init(Resources& res, RenderScene& rscene, cons
 
       shaderInfo = nvvkglsl::GlslCompiler::makeShaderModuleCreateInfo(m_shaders.computeGeometryBlasSharing);
       vkCreateComputePipelines(res.m_device, nullptr, 1, &compInfo, nullptr, &m_pipelines.computeGeometryBlasSharing);
+    }
+
+    if(config.useSeparateGroups)
+    {
+      shaderInfo = nvvkglsl::GlslCompiler::makeShaderModuleCreateInfo(m_shaders.computeTraversalGroups);
+      vkCreateComputePipelines(res.m_device, nullptr, 1, &compInfo, nullptr, &m_pipelines.computeTraversalGroups);
     }
   }
 
@@ -546,6 +572,18 @@ void RendererRayTraceClustersLod::render(VkCommandBuffer cmd, Resources& res, Re
 
     vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, m_pipelines.computeTraversalRun);
     vkCmdDispatch(cmd, getWorkGroupCount(frame.traversalPersistentThreads, TRAVERSAL_RUN_WORKGROUP), 1, 1);
+
+    if(m_config.useSeparateGroups)
+    {
+      memBarrier.srcAccessMask = VK_ACCESS_SHADER_WRITE_BIT;
+      memBarrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT | VK_ACCESS_UNIFORM_READ_BIT | VK_ACCESS_INDIRECT_COMMAND_READ_BIT;
+      vkCmdPipelineBarrier(cmd, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT,
+                           VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT | VK_PIPELINE_STAGE_DRAW_INDIRECT_BIT, 0, 1,
+                           &memBarrier, 0, nullptr, 0, nullptr);
+
+      vkCmdBindPipeline(cmd, VK_PIPELINE_BIND_POINT_COMPUTE, m_pipelines.computeTraversalGroups);
+      vkCmdDispatchIndirect(cmd, m_sceneBuildBuffer.buffer, offsetof(shaderio::SceneBuilding, indirectDispatchGroups));
+    }
 
     memBarrier.srcAccessMask = VK_ACCESS_SHADER_WRITE_BIT;
     memBarrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT | VK_ACCESS_UNIFORM_READ_BIT;